my-agents-resume/my-agents-resume/requirements.txt
fastapi>=0.115.5
uvicorn>=0.27.0
python-multipart>=0.0.6
pdfplumber>=0.10.0
python-docx>=1.1.0
google-adk>=1.16.0
google-genai>=1.45.0


